chore: adicionando menus genericos a sidebar
