+++ src/components/sidebar/Sidebar.tsx
@@ -1,0 +1,195 @@
+import React from 'react';
+import { Sidebar as LibSidebar, Menu, MenuItem, SubMenu, menuClasses, MenuItemStyles }  from 'react-pro-sidebar';
+import { GoGear, GoBell, GoSignOut, GoBook} from "react-icons/go";
+import { SidebarHeader } from './SidebarHeader';
+import { SidebarUser } from './SidebarUser';
+import { Badge } from './Badge';
+import { Typography } from './Typography';
+import { Link } from "react-router-dom";
+
+type Theme = 'light' | 'dark';
+
+const themes = {
+  light: {
+    LibSidebar: {
+      backgroundColor: '#ffffff',
+      color: '#607489',
+    },
+    menu: {
+      menuContent: '#fbfcfd',
+      icon: '#0098e5',
+      hover: {
+        backgroundColor: '#c5e4ff',
+        color: '#44596e',
+      },
+      disabled: {
+        color: '#9fb6cf',
+      },
+    },
+  },
+  dark: {
+    LibSidebar: {
+      backgroundColor: '#0b2948',
+      color: '#8ba1b7',
+    },
+    menu: {
+      menuContent: '#082440',
+      icon: '#59d0ff',
+      hover: {
+        backgroundColor: '#00458b',
+        color: '#b6c8d9',
+      },
+      disabled: {
+        color: '#3e5e7e',
+      },
+    },
+  },
+};
+
+// hex to rgba converter
+const hexToRgba = (hex: string, alpha: number) => {
+  const r = parseInt(hex.slice(1, 3), 16);
+  const g = parseInt(hex.slice(3, 5), 16);
+  const b = parseInt(hex.slice(5, 7), 16);
+
+  return `rgba(${r}, ${g}, ${b}, ${alpha})`;
+};
+
+export default function Sidebar(){
+  const [collapsed, setCollapsed] = React.useState(false);
+  const [toggled, setToggled] = React.useState(false);
+  const [broken, setBroken] = React.useState(false);
+  const [rtl, setRtl] = React.useState(false);
+  const [hasImage, setHasImage] = React.useState(false);
+  const [theme, setTheme] = React.useState<Theme>('light');
+
+  // handle on RTL change event
+  const handleRTLChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setRtl(e.target.checked);
+  };
+
+  // handle on theme change event
+  const handleThemeChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setTheme(e.target.checked ? 'dark' : 'light');
+  };
+
+  // handle on image change event
+  const handleImageChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setHasImage(e.target.checked);
+  };
+
+  const menuItemStyles: MenuItemStyles = {
+    root: {
+      fontSize: '13px',
+      fontWeight: 400,
+    },
+    icon: {
+      color: themes[theme].menu.icon,
+      [`&.${menuClasses.disabled}`]: {
+        color: themes[theme].menu.disabled.color,
+      },
+    },
+    SubMenuExpandIcon: {
+      color: '#b6b7b9',
+    },
+    subMenuContent: ({ level }) => ({
+      backgroundColor:
+        level === 0
+          ? hexToRgba(themes[theme].menu.menuContent, hasImage && !collapsed ? 0.4 : 1)
+          : 'transparent',
+    }),
+    button: {
+      [`&.${menuClasses.disabled}`]: {
+        color: themes[theme].menu.disabled.color,
+      },
+      '&:hover': {
+        backgroundColor: hexToRgba(themes[theme].menu.hover.backgroundColor, hasImage ? 0.8 : 1),
+        color: themes[theme].menu.hover.color,
+      },
+    },
+    label: ({ open }) => ({
+      fontWeight: open ? 600 : undefined,
+    }),
+  };
+
+  return (
+    <div style={{ display: 'flex', height: '100%', direction: rtl ? 'rtl' : 'ltr' }}>
+      <LibSidebar
+        collapsed={collapsed}
+        toggled={toggled}
+        onBackdropClick={() => setToggled(false)}
+        onBreakPoint={setBroken}
+        image="https://user-images.githubusercontent.com/25878302/144499035-2911184c-76d3-4611-86e7-bc4e8ff84ff5.jpg"
+        rtl={rtl}
+        breakPoint="md"
+        backgroundColor={hexToRgba(themes[theme].LibSidebar.backgroundColor, hasImage ? 0.9 : 1)}
+        rootStyles={{
+          color: themes[theme].LibSidebar.color,
+        }}
+      >
+        <div style={{ display: 'flex', flexDirection: 'column', height: '100%' }}>
+          <SidebarHeader rtl={rtl} collapsed={collapsed} setCollapsed={setCollapsed} style={{ marginBottom: '5px', marginTop: '16px' }} />
+          <SidebarUser rtl={rtl} collapsed={collapsed} setCollapsed={setCollapsed} style={{ marginBottom: '24px', marginTop: '5px' }} />
+          <div style={{ flex: 1, marginBottom: '32px' }}>
+            <Menu menuItemStyles={menuItemStyles}>
+              
+              <MenuItem 
+                icon={<GoBell />}
+                suffix={
+                  <Badge variant="danger" shape="circle">
+                    6
+                  </Badge>
+                }
+              >
+                <Link to="/notifications">Notificações</Link>
+              </MenuItem>
+
+              <MenuItem 
+                icon={<GoBook />}
+              >
+                <Link to="/services">Serviços</Link>
+              </MenuItem>
+              
+              <MenuItem 
+                icon={<GoGear />}
+              >
+                <Link to="/userconfig">Configurações</Link>
+              </MenuItem>
+
+              <MenuItem 
+                icon={<GoSignOut />}
+              >
+                <Link to="/login">Sair</Link>
+              </MenuItem>
+            </Menu>
+            <div style={{ padding: '0 24px', marginBottom: '8px', marginTop: '32px' }}>
+              <Typography
+                variant="body2"
+                fontWeight={600}
+                style={{ opacity: collapsed ? 0 : 0.7, letterSpacing: '0.5px' }}
+              >
+                Materias
+              </Typography>
+            </div>
+
+            <Menu menuItemStyles={menuItemStyles}>
+            <SubMenu
+                label="materia"
+                icon={""}
+                suffix={
+                  <Badge variant="danger" shape="circle">
+                    6
+                  </Badge>
+                }
+              >
+                <MenuItem> menu</MenuItem>
+                <MenuItem> menu</MenuItem>
+                <MenuItem> menu</MenuItem>
+              </SubMenu>
+            </Menu>
+          </div>
+        </div>
+      </LibSidebar>
+    </div>
+  );
+};
